#32639: Make IPTR unsigned long

+++ gcc/testsuite/g++.target/riscv/tt/sfpi/gtzero-26919-qsr.C
@@ -2,7 +2,7 @@
 // { dg-final { check-function-bodies "**" "" } }
 
 namespace ckernel{
-extern unsigned *instrn_buffer;
+extern unsigned long *instrn_buffer;
 }
 #include <sfpi.h>
